+++ backend/app/repositories/business/__init__.py
@@ -2,10 +2,12 @@ from .instrument_type_repository import InstrumentTypeRepository
 from .location_repository import LocationRepository
 from .manufacturer_repository import ManufacturerRepository
 from .material_repository import MaterialRepository
+from .instrument_repository import InstrumentRepository
 
 __all__ = [
     "InstrumentTypeRepository",
     "LocationRepository",
     "ManufacturerRepository",
     "MaterialRepository",
+    "InstrumentRepository",
 ]